refactor: removing all redux and related code
