+++ package.json
@@ -41,8 +41,6 @@
     "@fortawesome/react-fontawesome": "^0.2.0",
     "@openedx/frontend-plugin-framework": "^1.7.0",
     "@openedx/paragon": "^23.4.5",
-    "@redux-devtools/extension": "3.3.0",
-    "@reduxjs/toolkit": "^2.0.0",
     "@tanstack/react-query": "^5.90.16",
     "classnames": "^2.3.1",
     "core-js": "3.47.0",
@@ -55,14 +53,9 @@
     "react-dom": "^18.3.1",
     "react-helmet": "^6.1.0",
     "react-intl": "6.8.9",
-    "react-redux": "^7.2.4",
     "react-router-dom": "6.30.3",
     "react-share": "^4.4.0",
-    "redux": "4.2.1",
-    "redux-logger": "3.0.6",
-    "redux-thunk": "2.4.2",
     "regenerator-runtime": "^0.14.0",
-    "reselect": "^4.0.0",
     "universal-cookie": "^4.0.4",
     "util": "^0.12.4"
   },
@@ -80,7 +73,6 @@
     "jest-expect-message": "^1.1.3",
     "jest-when": "^3.6.0",
     "react-dev-utils": "^12.0.0",
-    "react-test-renderer": "^18.3.1",
-    "redux-mock-store": "^1.5.4"
+    "react-test-renderer": "^18.3.1"
   }
 }
