+++ src/containers/CourseCard/components/CourseCardBanners/CreditBanner/views/MustRequestContent.jsx
@@ -3,7 +3,7 @@ import PropTypes from 'prop-types';
 
 import { useIntl } from '@edx/frontend-platform/i18n';
 
-import { useIsMasquerading } from 'hooks/useIsMasquerading';
+import { useIsMasquerading } from 'hooks';
 import CreditContent from './components/CreditContent';
 import ProviderLink from './components/ProviderLink';
 import hooks from './hooks';
