+++ src/containers/CourseCard/components/CourseCardBanners/CreditBanner/views/PendingContent.test.jsx
@@ -1,17 +1,12 @@
 import { render, screen } from '@testing-library/react';
 import { IntlProvider } from '@edx/frontend-platform/i18n';
-import { useIsMasquerading } from 'hooks/useIsMasquerading';
-import { useCourseData } from 'hooks';
+import { useCourseData, useIsMasquerading } from 'hooks';
 
 import messages from './messages';
 import PendingContent from './PendingContent';
 
 jest.mock('hooks', () => ({
-  reduxHooks: { useCardCreditData: jest.fn(), useMasqueradeData: jest.fn() },
   useCourseData: jest.fn(),
-}));
-
-jest.mock('hooks/useIsMasquerading', () => ({
   useIsMasquerading: jest.fn(),
 }));
