--- src/data/redux/app/index.js
@@ -1,2 +1,0 @@
-export { actions, reducer } from './reducer';
-export { default as selectors } from './selectors';
