+++ src/hooks/index.js
@@ -1,13 +1,11 @@
-import * as redux from 'data/redux/hooks';
-import * as api from './api';
 import * as utils from './utils';
 import useCourseData from './useCourseData';
 import useCourseTrackingEvent from './useCourseTrackingEvent';
 import useEntitlementInfo from './useEntitlementInfo';
+import useIsMasquerading from './useIsMasquerading';
 
-export const reduxHooks = redux;
-export const apiHooks = api;
 export const utilHooks = utils;
 export { useCourseData };
 export { useCourseTrackingEvent };
 export { useEntitlementInfo };
+export { useIsMasquerading };
