--- src/test/utils.js
@@ -1,5 +1,0 @@
-// TODO: redux cleanup
-/* export const mockSuccess = (returnValFn) => (...args) => Promise.resolve(returnValFn(...args));
-
-export const mockFailure = (returnValFn) => (...args) => Promise.reject(returnValFn(...args));
-*/
